refactor: give combat/ foundry/ + logic/ folder parity (no logic class)
Mirror the combatant/ layout without a SohlLogic class — Combat has no
actions, per-combat state, or lifecycle to justify one.

- Move SohlCombat.ts (document + data model) to combat/foundry/.
- Move combat-logic.ts (pure resolveGroupSeeding planner) to combat/logic/.
- Add combat/logic/** to the ESLint Foundry-free zone and the purity smoke
  test, enforcing its purity.
- Update importers (SohlSystem, SohlCombatant, the seeding test) and annotate
  one map callback that lost inference after the move.

No behavior change.
