+++ eslint.config.js
@@ -61,6 +61,7 @@ const FOUNDRY_FREE_ZONES = [
     "src/document/item/logic/**/*.ts",
     "src/document/actor/logic/**/*.ts",
     "src/document/combatant/logic/**/*.ts",
+    "src/document/combat/logic/**/*.ts",
     "src/domain/**/*.ts",
     "src/core/SohlLogic.ts",
     "src/core/SohlActionContext.ts",
