+++ src/document/combat/logic/combat-logic.ts
